refactor: 잘못된 모듈명 수정
